Extract requestAction() into a trait.
Having a separate trait allows us to reduce the number of places that
requestAction is exposed, and also allow people to use requestAction
where they need without inheritance.
